Split UpdateAlg so that it can be used in SbtAlg
See https://github.com/fthomas/scala-steward/pull/1144 for context.
